+++ core/Cargo.toml
@@ -25,6 +25,7 @@ hashbrown.workspace = true
 enum-iterator.workspace = true
 byteorder.workspace = true
 num-traits.workspace = true
+paste.workspace = true
 
 # Optional dependencies
 serde = { workspace = true, features = ["derive"], optional = true }
